Feature#63

+++ SpaceZBE/src/main/java/com/spaceZ/member/MemberDAO.java
@@ -5,7 +5,7 @@ import javax.mail.MessagingException;
 public interface MemberDAO {
 	
 	public int signUp(MemberVO vo);
-	public int login(MemberVO vo);
+	public long login(MemberVO vo);
 	public int idCheck(MemberVO vo);
 	public int sendEmail(EmailVO vo) throws MessagingException;
 	public int update(MemberVO vo);
